+++ Backend/Lyzer/Common/DTO/OverviewDataDTO.cs
@@ -3,6 +3,7 @@
     public class OverviewDataDTO
     {
         public required RaceWeekendProgressDTO RaceWeekendProgress { get; set; }
+        public required UpcomingRaceWeekendDTO UpcomingRaceWeekend { get; set; }
     }
 
     public class RaceWeekendProgressDTO
@@ -10,5 +11,13 @@
         public required string Name { get; set; }
         public required bool Ongoing { get; set; }
         public required int WeekendProgress { get; set; }
+    }
+
+    public class UpcomingRaceWeekendDTO
+    {
+        public required bool IsRaceWeekend { get; set; }
+        public required int TimeToRaceWeekendProgress { get; set; }
+        public required string Status { get; set; }
+        public required double TimeToRaceWeekend { get; set; }
     }
 }
